Create README.md
Embark on a data-driven journey through the bustling streets of Seoul! This project taps into the city's vibrant heartbeat by predicting bike rental demands with precision. Utilizing a rich dataset from 2017-2018, I navigated through complex patterns of weather variations, holiday impacts, and hourly usage trends.

The adventure involved harnessing the power of Linear and Support Vector Regression models, fine-tuned with hyperparameter optimization to achieve a marked improvement in prediction accuracy. My passion for data storytelling is reflected in the detailed visualizations and insightful analyses that make urban mobility patterns understandable and actionable.

This project was a thrilling blend of challenge and creativity, providing me with a profound appreciation for the nuances of machine learning in real-world applications. It demonstrated the potential of predictive analytics to revolutionize urban planning and resource management in metropolitan areas.